[github-action] add cache, and temp use own runner.

+++ scripts/install-gnu-bison.sh
@@ -1,8 +1,16 @@
-mkdir bison && cd bison
-curl https://ftp.gnu.org/gnu/bison/bison-3.8.2.tar.xz -o bison.tar.xz
-tar -xvJf bison.tar.xz
-rm bison.tar.xz
-cd bison-3.8.2
-./configure
+# if bison folder exists, skip building.
+if [ ! -d bison ]
+then
+	mkdir bison
 
-make -j$(nproc)
+	cd bison
+
+	curl https://ftp.gnu.org/gnu/bison/bison-3.8.2.tar.xz -o bison.tar.xz
+	tar -xvJf bison.tar.xz
+	rm bison.tar.xz
+
+	cd bison-3.8.2
+	./configure
+
+	make -j$(nproc)
+fi
